// client/pokemonapi/model/pokemonDetails.ts
/**
 * Pokemon Powers API
 * Add your description
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * PokemonDetails
 */
export interface PokemonDetails { 
    /**
     * The color of the Pokemon
     */
    color?: string;
    /**
     * The weight of the Pokemon
     */
    weight?: string;
    /**
     * The height of the Pokemon
     */
    height?: string;
    /**
     * The type of the Pokemon
     */
    type?: string;
    /**
     * The category of the Pokemon
     */
    category?: string;
    /**
     * Games the Pokemon has appeared in
     */
    games?: Array<string>;
    /**
     * TV shows the Pokemon has appeared in
     */
    tvShows?: Array<string>;
    /**
     * Books the Pokemon has appeared in
     */
    books?: Array<string>;
    /**
     * Posters featuring the Pokemon
     */
    posters?: Array<string>;
}